src/test/java/edu/tntech/csc2310/CourseCatalogTest.java
package edu.tntech.csc2310;

import org.junit.BeforeClass;
import org.junit.Test;

import java.util.ArrayList;

import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNull;

public class CourseCatalogTest {

    public static CourseCatalog catalog;
    private static Object IOException;

    long timeNow = System.currentTimeMillis();

    @BeforeClass
    public static void setUp() throws Exception {
        CourseCatalogTest.catalog = new CourseCatalog("CSC", "202180", IOException);
    }

    public static void setIOException(Object IOException) {
        CourseCatalogTest.IOException = IOException;
    }

    @Test
    public void getCourse() {
        Course c = CourseCatalogTest.catalog.getCourse("2770");
        assertEquals("Course test", "Intro to Systems & Networking", c.getTitle());
        assertEquals("Course test credits", 3, c.getCredits());
    }

    @Test
    public void getCourseNonExistent() {
        Course c1 = CourseCatalogTest.catalog.getCourse("2001");
        assertNull(c1);
    }

    @Test
    public void badSubject(){
        CourseCatalog c = new CourseCatalog("Missing", "202180", IOException);
        assertNull(c.getSubject());
        assertNull(c.getCatalogYear());
    }

    @Test
    public void trimSubject(){
        CourseCatalog c = new CourseCatalog(" CSC ", "202180", IOException);
        assertEquals("subject trim", "CSC", c.getSubject());
    }

    @Test
    public void toUpperSubject(){
        CourseCatalog c = new CourseCatalog("math", "202180", IOException);
        assertEquals("subject lowercase", "MATH", c.getSubject());
    }


    @Test
    public void getCourseNumbers() {
        ArrayList list = CourseCatalog.getCourseNumbers("CSC", "202180");
        assertEquals("CourseNumber test", 62, list.size());
        list = CourseCatalog.getCourseNumbers("COMM", "202180");
        assertEquals("CourseNumber test", 37, list.size());
    }

    @Test
    public void getCatalogYear() {
        assertEquals("Catalog Year", "202180", CourseCatalogTest.catalog.getCatalogYear());
    }

    @Test
    public void getSubject() {
        assertEquals("Subject", "CSC", CourseCatalogTest.catalog.getSubject());
    }
}